# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servidor API para el Proyecto Sakoo (Arquitectura Limpia).",
        "title": "Sakoo Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte Sakoo",
            "url": "http://www.sakoo.com",
            "email": "soporte@sakoo.com"
        },
        "license": {
            "name": "Propietaria",
            "url": "http://www.sakoo.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/admin/scrape-mercantil": {
            "post": {
                "description": "Ejecuta de forma síncrona el web scraper de Mercantil para actualizar las tasas de cambio de este banco.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Administración"
                ],
                "summary": "Forzar raspado de tasas Mercantil",
                "responses": {
                    "200": {
                        "description": "Raspado y actualización de Mercantil ejecutados con éxito",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Error interno al realizar el scraping",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/api/admin/scrape-now": {
            "post": {
                "description": "Ejecuta de forma síncrona el web scraper del Banco Central de Venezuela (BCV) para actualizar las tasas de cambio.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Administración"
                ],
                "summary": "Forzar raspado de tasas BCV",
                "responses": {
                    "200": {
                        "description": "Raspado y actualización del BCV ejecutados con éxito",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Error interno al realizar el scraping",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/rates/calculate": {
            "post": {
                "description": "Calcula la conversión de un monto específico a la moneda objetivo según tasas vigentes o históricas.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core Business"
                ],
                "summary": "Calcular conversión de monedas",
                "parameters": [
                    {
                        "description": "Datos para la conversión",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.ConversionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moneda no encontrada o monto negativo",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/rates/dashboard": {
            "get": {
                "description": "Retorna la última tasa de la moneda especificada, su porcentaje de variación y su historial de tasas recientes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core Business"
                ],
                "summary": "Obtener resumen del dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código de la moneda (ej. USD)",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moneda no encontrada o error interno",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "track_code": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.ConversionRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "date": {
                    "description": "Opcional para calculadora retroactiva",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Coloque el token JWT de la siguiente manera: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}